Split canvas pane focus test spy
